Fix compile error and rename file
